% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{key_from_passphrase}
\alias{key_from_passphrase}
\title{Derive a secret key from a passphrase}
\usage{
key_from_passphrase(passphrase, bytes = 32L)
}
\arguments{
\item{passphrase}{Character string (e.g. password).}

\item{bytes}{Length of derived key in bytes (default 32).}
}
\value{
Raw vector of derived key bytes.
}
\description{
Derive a secret key from a passphrase
}
\examples{
k <- key_from_passphrase("my-secret")
}
